# services/agentiq-wallet/openapi.yaml
openapi: 3.0.3
info:
  title: AgentiQ Wallet Service API
  version: 0.1.0
servers:
  - url: http://localhost:8090
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: DID-JWS
  parameters:
    IdentityStateHeader:
      in: header
      name: x-identity-state
      required: false
      schema:
        type: string
        enum: [anonymous, persona, root, kybe]
      description: Desired identity level for this request.
paths:
...
  /x402/remote/session:
    post:
      summary: Create Remote Custody session
      x-require-level: persona
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ttlSec: { type: integer, minimum: 1, default: 900 }
                caps:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  did: { type: string }
                  persona: { type: string }
                  expiresIn: { type: integer }
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdentityStateHeader'
...